exercises blog 4.15 to 4.23 done

+++ part4/controllers/blogsController.js
@@ -11,22 +11,19 @@ blogRouter.get('/', async (request, response) => {
 })
 
 blogRouter.post('/', userExtractor, async (request, response) => {
-    if(!request.body) return response.status(400).end()
+    const { title, author, url, likes } = request.body
 
-    const { title, author, url, likes, userId } = request.body
-
-    if(!userId) return response.status(400).json({ error: 'request must contain userId field' })
+    if(!title || !url ) return response.status(400).json({ error: 'title and url are required parameters' })
 
     const user = await User.findById(request.user.id)
-
-    if(!user) return response.status(400).json({ error: 'Invalid or missing id' })
+    if(!user) return response.status(401).json({ error: 'Invalid or missing user' })
 
     const blogToAdd = new Blog({
         title,
         author,
         url,
         likes,
-        user: userId
+        user: user._id
     })
 
     const addedBlog = await blogToAdd.save()
@@ -39,8 +36,6 @@ blogRouter.post('/', userExtractor, async (request, response) => {
 })
 
 blogRouter.delete('/:id', userExtractor, async (request, response) => {
-    if(!request.params) return response.status(400).end()
-
     const id = request.params.id
 
     const blogToDelete = await Blog.findById(id)
@@ -49,19 +44,31 @@ blogRouter.delete('/:id', userExtractor, async (request, response) => {
         return response.status(401).json({ error: 'blog can be deleted only by the owner' })
     }
 
+    const user = await User.findById(request.user.id.toString())
+    if(!user) return response.status(401).json({ error: 'Invalid or missing user' })
+
     await Blog.deleteOne({ id: id })
 
     response.status(204).end()
 })
 
-blogRouter.put('/:id', async (request, response) => {
-    if(!request.params) return response.status(400).end()
-    if(!request.body) return response.status(400).end()
+blogRouter.put('/:id', userExtractor, async (request, response) => {
+    const user = await User.findById(request.user.id.toString())
+    if(!user) return response.status(401).json({ error: 'Invalid or missing user' })
 
     const id = request.params.id
-    const blogToUpdate = request.body
+    const newBlogInfo = request.body
 
-    const updatedBlog = await Blog.findByIdAndUpdate(id, blogToUpdate, { new: true, runValidators: true, context: 'query' })
+    const blogToUpdate = await Blog.findById(id)
+    const blogOwner = blogToUpdate.user.toJSON()
+
+    const receivedUserId = request.user.id
+
+    if(receivedUserId !== blogOwner) {
+        return response.status(401).json({ error: 'blog can be updated only by the owner' })
+    }
+
+    const updatedBlog = await Blog.findByIdAndUpdate(id, newBlogInfo, { new: true, runValidators: true, context: 'query' })
 
     response.status(200).json(updatedBlog)
 })
